// packages/core/src/squads/generated/accounts/multisig.ts
export const MULTISIG_DISCRIMINATOR: ReadonlyUint8Array = new Uint8Array([
  224, 116, 121, 186, 68, 161, 79, 236,
]);

export function getMultisigDiscriminatorBytes(): ReadonlyUint8Array {
  return fixEncoderSize(getBytesEncoder(), 8).encode(MULTISIG_DISCRIMINATOR);
}

export type Multisig = {
  discriminator: ReadonlyUint8Array;
  /** Key that is used to seed the multisig PDA. */
  createKey: Address;
  /**
   * The authority that can change the multisig config.
   * This is a very important parameter as this authority can change the members and threshold.
   *
   * The convention is to set this to `Pubkey::default()`.
   * In this case, the multisig becomes autonomous, so every config change goes through
   * the normal process of voting by the members.
   *
   * However, if this parameter is set to any other key, all the config changes for this multisig
   * will need to be signed by the `config_authority`. We call such a multisig a "controlled multisig".
   */
  configAuthority: Address;
  /** Threshold for signatures. */
  threshold: number;
  /** How many seconds must pass between transaction voting settlement and execution. */
  timeLock: number;
  /** Last transaction index. 0 means no transactions have been created. */
  transactionIndex: bigint;
  /**
   * Last stale transaction index. All transactions up until this index are stale.
   * This index is updated when multisig config (members/threshold/time_lock) changes.
   */
  staleTransactionIndex: bigint;
  /**
   * The address where the rent for the accounts related to executed, rejected, or cancelled
   * transactions can be reclaimed. If set to `None`, the rent reclamation feature is turned off.
   */
  rentCollector: Option<Address>;
  /** Bump for the multisig PDA seed. */
  bump: number;
  /** Members of the multisig. */
  members: Array<Member>;
};