// packages/core/src/squads/generated/accounts/proposal.ts
export type Proposal = {
  discriminator: ReadonlyUint8Array;
  /** The multisig this belongs to. */
  multisig: Address;
  /** Index of the multisig transaction this proposal is associated with. */
  transactionIndex: bigint;
  /** The status of the transaction. */
  status: ProposalStatus;
  /** PDA bump. */
  bump: number;
  /** Keys that have approved/signed. */
  approved: Array<Address>;
  /** Keys that have rejected. */
  rejected: Array<Address>;
  /** Keys that have cancelled (Approved only). */
  cancelled: Array<Address>;
};