// packages/core/src/squads/generated/types/proposalStatus.ts
export type ProposalStatus =
  | { __kind: "Draft"; timestamp: bigint }
  | { __kind: "Active"; timestamp: bigint }
  | { __kind: "Rejected"; timestamp: bigint }
  | { __kind: "Approved"; timestamp: bigint }
  | { __kind: "Executing" }
  | { __kind: "Executed"; timestamp: bigint }
  | { __kind: "Cancelled"; timestamp: bigint };